requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Data processing
pandas>=2.2.0  # Updated for Python 3.13 and NumPy 2.x compatibility
numpy>=2.0.0  # Updated for Python 3.13 compatibility
openpyxl==3.1.2  # For Excel file reading

# AI/ML
google-generativeai>=0.3.0  # Google Gemini API

# CORS is handled by FastAPI's built-in CORSMiddleware

# Optional: For production database (PostgreSQL)
# psycopg2-binary==2.9.9

# Optional: For email notifications
# aiosmtplib==3.0.1

# Optional: For advanced ML (if needed)
# scikit-learn==1.3.2
# scipy==1.11.4
python-dotenv==1.0.1
email-validator==2.1.0.post1
bcrypt==4.1.2
gunicorn==21.2.0
typing-extensions>=4.9.0
httpx==0.26.0

